pipeline/annotation/requirements.txt
# =============================================================================
# PROJET GAZA - DÉPENDANCES PYTHON
# =============================================================================

# --- Data Processing ---
pandas>=2.0.0
polars>=0.20.0
pyarrow>=14.0.0
duckdb>=0.9.0

# --- NLP & Text Processing ---
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# Pre-trained models
# pip install camembert-base (via transformers)

# --- Topic Modeling ---
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.5
hdbscan>=0.8.33

# --- Text preprocessing ---
spacy>=3.7.0
# python -m spacy download fr_core_news_lg
unidecode>=1.3.0
emoji>=2.10.0
regex>=2023.0.0

# --- Change Point Detection ---
ruptures>=1.1.9

# --- Statistics ---
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
altair>=5.2.0

# --- Dashboard ---
streamlit>=1.29.0

# --- Annotation ---
openai>=1.40.0
# Label Studio (optional, run separately)
# pip install label-studio

# --- Inter-annotator Agreement ---
krippendorff>=0.6.0

# --- Utilities ---
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0
requests>=2.31.0

# --- Jupyter ---
jupyter>=1.0.0
ipywidgets>=8.1.0

# --- Development ---
black>=23.0.0
isort>=5.13.0
pytest>=7.4.0


